# openapi/cost.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OTel AI Cost Local Serve API",
...
    "version": "0.1.0",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8792",
      "description": "Local MVP default"
    }
  ],
...
  "paths": {
...
    "/report.json": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Structured cost report JSON",
        "description": "Serve default is UTC daily rollup (`groupBy=day`) including `totalUsd` and `days[]`. CORS OPTIONS unlisted Origin → **403** `cors_denied`. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`. Includes `byTenant` and `budgetBreaches` (empty array when no `--tenant-budget` / env).",
        "operationId": "getReport",
        "security": [],
        "responses": {
          "200": {
            "description": "Cost snapshot JSON",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReport"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      }
    },
    "/v1/costs.csv": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Finance CSV cost export",
        "description": "Daily-by-model+tenant totals as CSV with header `date,model,spanCount,usd,tenant` (`tenant` last so `date,model,spanCount,usd` greps still match) plus a TOTAL row. Same cost model as `/report.json` (`day` → `date`, per-model `spanCount`, `usd`). Missing/empty span attr `tenant` → `_`. Empty spans → header-only **200**. CORS + `X-Request-Id` same as other GET. Access-logged (not skipped like `/metrics`). Alias: `GET /v1/costs?format=csv`. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`.",
        "operationId": "getCostsCsv",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "CSV cost snapshot (header + day/model rows + TOTAL; header only when empty)",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                },
                "example": "date,model,spanCount,usd,tenant\n2024-08-11,gpt-4o-mini,1,0.000450,acme\nTOTAL,,1,0.000450,\n"
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/costs.md": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "FinOps Markdown cost report",
        "description": "GitHub-flavored Markdown for Slack / email / GitHub Actions `$GITHUB_STEP_SUMMARY`. Shape: `# otel-ai-cost`, `**totalUsd:**` / `**spans:**`, `## by model` table (`model | usd | spans`), `## by tenant` table (`tenant | usd | spans`). `|` in cells is escaped. Empty spans → heading + zeros + table headers (no data rows), **200**. Same cost model as `/report.json`. CORS + `X-Request-Id` same as other GET. Access-logged (not skipped like `/metrics`). Alias: `GET /v1/costs?format=md`. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`.",
        "operationId": "getCostsMd",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Markdown cost snapshot (heading + totals + by-model / by-tenant tables; heading + zeros when empty)",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                },
                "example": "# otel-ai-cost\n**totalUsd:** 0.000450  **spans:** 1\n## by model\n| model | usd | spans |\n| --- | --- | --- |\n| gpt-4o-mini | 0.000450 | 1 |\n## by tenant\n| tenant | usd | spans |\n| --- | --- | --- |\n| acme | 0.000450 | 1 |\n"
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/costs.gha.txt": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "GitHub Actions annotations alias (same as GET /v1/costs?format=gha)",
        "description": "Same body as `GET /v1/costs?format=gha` (workflow commands: global `::error title=budget::totalUsd X > budget Y` when `--budget` maxTotalUsd is exceeded; each tenant `::error title=tenant/<id>::usd X > budget Y` from `budgetBreaches`; no breach → empty body 200, no `::error`). Content-Type `text/plain`. Does not require GitHub. CORS OPTIONS unlisted Origin → **403** `cors_denied`. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`.",
        "operationId": "getCostsGha",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "GHA workflow commands (empty body is still 200)",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "::error title=tenant/acme::usd 0.0125 > budget 0.0001\n"
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/costs": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Cost snapshot (json | csv | md | gha)",
        "description": "Versioned cost snapshot. Query `format=json` (default) is the same DailyReport body as `GET /report.json`. `format=csv` is the same `text/csv` body as `GET /v1/costs.csv`. `format=md` is the same `text/markdown` body as `GET /v1/costs.md`. `format=gha` (alias `annotations`) is GitHub Actions workflow commands (`text/plain`; same as `GET /v1/costs.gha.txt`; global `::error title=budget::totalUsd X > budget Y`; tenant `::error title=tenant/<id>::usd X > budget Y`; no breach → empty 200). Unknown format → **400** `{error:\"bad_format\"}`. Existing `/report.json` clients are unchanged. CORS OPTIONS unlisted Origin → **403** `cors_denied`. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`. JSON includes `budgetBreaches` (same as `/report.json`).",
        "operationId": "getCosts",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/CostFormat"
          }
        ],
        "responses": {
          "200": {
            "description": "Cost snapshot in the requested format",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReport"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                },
                "example": "date,model,spanCount,usd,tenant\n2024-08-11,gpt-4o-mini,1,0.000450,acme\nTOTAL,,1,0.000450,\n"
              },
              "text/markdown": {
                "schema": {
                  "type": "string"
                },
                "example": "# otel-ai-cost\n**totalUsd:** 0.000450  **spans:** 1\n## by model\n| model | usd | spans |\n| --- | --- | --- |\n| gpt-4o-mini | 0.000450 | 1 |\n## by tenant\n| tenant | usd | spans |\n| --- | --- | --- |\n| acme | 0.000450 | 1 |\n"
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "::error title=tenant/acme::usd 0.0125 > budget 0.0001\n"
              }
            }
          },
          "400": {
            "description": "Unknown `format` query value. Body `{error:\"bad_format\"}`.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "bad_format",
                  "allowed": [
                    "csv",
                    "json",
                    "md",
                    "gha"
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/budgets": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Configured budget thresholds (not spend)",
        "description": "Returns configured global + tenant budget **thresholds**, not spend. `{ok:true, globalUsd: number|null, tenants:{acme:10,…}}`. `globalUsd` is `--budget` policy `maxTotalUsd` (missing/empty → `null`). `tenants` is `--tenant-budget` / env `OTEL_AI_COST_TENANT_BUDGETS` (alias `TENANT_BUDGETS`) map (missing → `{}`). Never includes spend, `budgetBreaches`, webhook secrets, or ingest tokens. CORS + `X-Request-Id` same as other GET. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`.",
        "operationId": "getBudgets",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Configured thresholds. Missing config → `globalUsd: null` and `tenants: {}`.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budgets"
                },
                "example": {
                  "ok": true,
                  "globalUsd": null,
                  "tenants": {
                    "acme": 10
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Pricing catalog (rates, not spend)",
        "description": "Returns the loaded price table as a catalog (rates, not spend). `{ok:true, models:[{id, inputPerMTok, outputPerMTok, …}], defaultModel:null, pack:null}`. Built-in table is `DEFAULT_PRICES` (`gpt-4o`, `gpt-4o-mini`, `claude-sonnet` with `inputPerMTok` / `outputPerMTok`). Field names match the table — do not invent rates. A single `usdPer1k` is used when that is all an entry has. Cached rates only when present. Empty/default still **200** with the built-in table (not necessarily empty). Never includes spend, webhook secrets, or ingest tokens. CORS + `X-Request-Id` same as other GET. Rate limit exceeded → **429** `{ok:false, reason:\"rate_limited\"}` + `Retry-After`.",
        "operationId": "getModels",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Pricing catalog from the loaded table. Default built-in pack is non-empty.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/AccessControlAllowOrigin"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Models"
                },
                "example": {
                  "ok": true,
                  "models": [
                    {
                      "id": "gpt-4o",
                      "inputPerMTok": 2.5,
                      "outputPerMTok": 10
                    }
                  ],
                  "defaultModel": null,
                  "pack": null
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/CorsDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "AccessControlAllowOrigin": {
        "description": "Echoed when request Origin matches `--cors-origins` / `OTEL_AI_COST_CORS_ORIGINS` (or `*` when configured). Absent when CORS disabled or origin not allowed.",
        "schema": {
          "type": "string",
          "example": "http://localhost:3000"
        }
      },
      "XRequestId": {
        "description": "Echo of incoming `X-Request-Id`, or a generated UUID when the request header is omitted/empty. Present on every response (incl 4xx / OPTIONS).",
        "schema": {
          "type": "string",
          "example": "mvp-req-id-a1b2c3d4"
        }
      },
...
    },
...
    "responses": {
      "CorsDenied": {
        "description": "CORS preflight (OPTIONS) for an Origin not in `--cors-origins` / `OTEL_AI_COST_CORS_ORIGINS`. Body `{error:\"forbidden\",reason:\"cors_denied\"}`. No `Access-Control-Allow-Origin`. Does not apply when CORS is disabled (then OPTIONS is 404 without ACAO). GET from a disallowed Origin still execute (e.g. 200) but omit ACAO.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "forbidden",
              "reason": "cors_denied"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "IngestBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Optional ingest token. Required only when INGEST_TOKEN / --ingest-token is set. Default off (local / stack-demo)."
      }
    }
  }
}